+++ pyrit/backend/models/scenarios.py
@@ -14,6 +14,7 @@ from typing import Any, Optional
 
 from pydantic import BaseModel, Field
 
+from pyrit.backend.models.attacks import AttackResultDetail
 from pyrit.backend.models.common import PaginationInfo
 
 
@@ -120,21 +121,6 @@ class ScenarioRunListResponse(BaseModel):
 # ============================================================================
 # Scenario Results Detail Models
 # ============================================================================
-
-
-class AttackResultDetail(BaseModel):
-    """Detailed result of a single attack within a scenario."""
-
-    attack_result_id: str = Field(..., description="Unique ID of this attack result")
-    conversation_id: str = Field(..., description="Conversation ID that produced this result")
-    objective: str = Field(..., description="Natural-language description of the attacker's objective")
-    outcome: str = Field(..., description="Attack outcome: success, failure, or undetermined")
-    outcome_reason: str | None = Field(None, description="Reason for the outcome")
-    last_response: str | None = Field(None, description="Model response from the final turn")
-    score_value: str | None = Field(None, description="Score value from the objective scorer")
-    executed_turns: int = Field(0, ge=0, description="Number of turns executed")
-    execution_time_ms: int = Field(0, ge=0, description="Execution time in milliseconds")
-    timestamp: datetime | None = Field(None, description="When the result was created")
 
 
 class AtomicAttackResults(BaseModel):
